feat: add 'single_step' mode
This is an optional mode where the version number will only be incremented by the minimum possible amount, regardless of how many underlying commits there were.
For example, assume the last release was v1.2.3 and your commit log is as follows:

  fix: ...
  feat: ...
  fix: ...
  fix: ...

If single_step is false (the default), this will result in the generated next version v1.3.2.
Setting single_step to true, however, will result in the version v1.3.0.