feat(#11): add `fold` option to `analyze_commits`
As noted in #11, for some users that are used to
[semantic-release](https://github.com/xotahal/fastlane-plugin-semantic_release/issues/11)
it is quite unexpected that the version number increments _for each commit_ rather than
_per changeset_.

This commit introduces the more expected behavior as a non-breaking change via the optional
`fold` parameter.

I have not added new tests at this time as implementing this has taken
way longer and grown more complex than I intended.